% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_Weibull_MLEs.R
\name{add_Weibull_mle}
\alias{add_Weibull_mle}
\title{Compute The MLEs of the Weibull Censored Data}
\usage{
add_Weibull_mle(censor_data)
}
\arguments{
\item{censor_data}{The value of function \code{simulate_weibull_data()}.}
}
\description{
Find out the MLEs for censored data with \code{survival} package.
}
\examples{
censor_data <- simulate_weibull_data(1, 5, 0.2, 1, 1)
( censor_data_with_mle <- add_Weibull_mle(censor_data) )
}
