% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_Weibull_MLEs.R
\name{add_Weibull_mle_robust}
\alias{add_Weibull_mle_robust}
\title{A Robust Algorithm to Compute The MLEs of the Weighted Weibull Censored Data}
\usage{
add_Weibull_mle_robust(censor_data, weight = "no")
}
\arguments{
\item{censor_data}{The value of function \code{simulate_weibull_data()}.}

\item{weight}{Specify the way to put a weight on the data. "no": do not put the weight on the data.}
}
\description{
Compute the MLEs of censored Weibull data by reparameterization and using \code{optim()} function.
}
\examples{
censor_data <- simulate_weibull_data(1, 5, 0.2, 1, 1)
( censor_data_with_mle <- add_Weibull_mle_robust(censor_data) )
}
