% Generated by roxygen2 (4.0.2): do not edit by hand
\name{capwords}
\alias{capwords}
\title{Capitalize every first word letter of a string}
\usage{
capwords(string, strict = FALSE)
}
\arguments{
\item{string}{a vector of string of characters.}

\item{strict}{logical. If \code{FALSE} (default) the case of the letters inside the words is not modified. If \code{TRUE}, all letters inside a word are lower case.}
}
\description{
Split the string into words (must be separated by a signle space \code{" "}) and capitalize every word.
}

