% Generated by roxygen2 (4.0.1): do not edit by hand
\name{diff_df}
\alias{diff_df}
\title{Extract rows that differ between two data frames}
\usage{
diff_df(df1, df2)
}
\arguments{
\item{df1,df2}{data frames}
}
\description{
This function generalise the concept of \code{setdiff} to \code{data.frame}. Only rows of the columns present in both data frames are compared.
}

