% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract_by_prefix}
\alias{extract_by_prefix}
\title{Extract from a data frame using prefix}
\usage{
extract_by_prefix(df, prefix, keep = NULL)
}
\arguments{
\item{df}{a data frame}

\item{prefix}{character, prefix of the variables to be extracted}

\item{keep}{character vector, specify variables without \code{prefix} to be returned}
}
\value{
a \code{data.frame} with the variables renamed without the \code{prefix}.
}
\description{
Extract only variables with a given \code{prefix}, allow for keeping variables without \code{prefix} by passing their names in \code{keep}.
}

