% Generated by roxygen2 (4.0.2): do not edit by hand
\name{facet_wrap_date_waiver}
\alias{facet_wrap_date_waiver}
\title{Date axis in selected panels of a facet wrap}
\usage{
facet_wrap_date_waiver(gplot, panel, date_0, format = "\%b \%d",
  n_ticks_max = NULL)
}
\arguments{
\item{gplot}{a \code{ggplot} or \code{grob} object.}

\item{panel}{numeric vector. Which panels should have date axis. Panels are numeroted from left to right and top to bottom.}

\item{date_0}{character. Date corresponding to the value 0 of the axis.}

\item{format}{character. The default for the format methods is \code{"\%b \%d"}, see \code{\link{strptime}} for other suitable \code{format}.}

\item{n_ticks_max}{numeric. Maximum number of ticks on the axis. This is useful when the original axis has too many ticks and date labels overlap.}
}
\description{
This function allows to transform the numeric axis labels of some panels into date fomat.
}

