% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_Weibull_MLEs.R
\name{find_Weibull_MLEs}
\alias{find_Weibull_MLEs}
\title{A Wrapper Function of \code{simulate_weibull_data}}
\usage{
find_Weibull_MLEs(censor_data)
}
\arguments{
\item{censor_data}{The value of function\code{simulate_weibull_data()}.}
}
\description{
This function only gives the MLEs, discarding the censoring data.
}
