% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_Weibull_MLEs.R
\name{find_Weibull_MLEs_robust}
\alias{find_Weibull_MLEs_robust}
\title{A Wrapper Function of \code{add_Weibull_mle_robust}}
\usage{
find_Weibull_MLEs_robust(censor_data)
}
\arguments{
\item{censor_data}{The value of function\code{add_Weibull_mle_robust()}.}
}
\description{
This function only gives the MLEs, discarding the censoring data.
}
