% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ggletter}
\alias{ggletter}
\title{Add letter}
\usage{
ggletter(letter, margin = c(0, 0.2, 0, 0), size = 20, hjust = -0.075,
  vjust = -0.1)
}
\arguments{
\item{letter}{character}

\item{margin}{vector of size 4}

\item{size}{text size (in pts)}

\item{hjust}{horizontal justification (in [0, 1])}

\item{vjust}{vertical justification (in [0, 1])}
}
\description{
Add letter to a ggplot
}

