% Generated by roxygen2 (4.0.2): do not edit by hand
\name{intersect_rec}
\alias{intersect_rec}
\title{Intersect a list of vectors}
\usage{
intersect_rec(.list)
}
\arguments{
\item{.list}{List of vectors to be intersected.}
}
\description{
Intersect a list of vectors recursively using \code{\link{intersect}}
}

