% Generated by roxygen2 (4.0.2): do not edit by hand
\name{merge_rec}
\alias{merge_rec}
\title{Merge a list of objects}
\usage{
merge_rec(.list, ...)
}
\arguments{
\item{.list}{List of objects (usually \code{data.frame}) to be merged}

\item{...}{Further arguments to be passed to \code{\link{merge}}}
}
\description{
Merge a list of objects recursively using \code{\link{merge}}
}

