% Generated by roxygen2 (4.0.2): do not edit by hand
\name{my_charmatch}
\alias{my_charmatch}
\title{String matching}
\usage{
my_charmatch(x, match)
}
\arguments{
\item{x}{vector of characters.}

\item{match}{vector of characters.}
}
\description{
Match each element of \code{x} to an element of \code{match} if either one of these element is a sub-string of the other. This is done by combining \code{\link{charmatch}} and \code{\link[stringr]{str_detect}}.
}
\seealso{
\code{\link{charmatch}}, \code{\link[stringr]{str_detect}}
}

