% Generated by roxygen2 (4.0.1): do not edit by hand
\name{NA omit and summarize}
\alias{NA omit and summarize}
\alias{any_na_omit}
\alias{last_na_omit}
\title{Basic accounting for NA values}
\usage{
last_na_omit(x)

any_na_omit(x)
}
\arguments{
\item{x}{vector}
}
\description{
All these functions performs the same as the basic version but first remove all \code{NA} elements. If all elements are \code{NA} then \code{NA} is returned, which is not always the case for the basic function.
}

