% Generated by roxygen2 (4.0.1): do not edit by hand
\name{reduce_duplicated_char}
\alias{reduce_duplicated_char}
\title{Reduce duplicated characters in a string}
\usage{
reduce_duplicated_char(x, char = "_")
}
\arguments{
\item{x}{Vector of strings}

\item{char}{Vector of characters}
}
\description{
This function reduces several consecutive occurences of \code{char} in \code{x} to a single occurence of \code{char}. Vectorised over \code{x} and \code{char}.
}
\examples{
x <- c("toto_1__2___3....4","toto..2___3")
char <- c("_",".")
reduce_duplicated_char(x,char)
}

