% Generated by roxygen2 (4.0.2): do not edit by hand
\name{remove_char}
\alias{remove_char}
\title{Remove characters from a string}
\usage{
remove_char(x, char = c(0:10, " ", "_"))
}
\arguments{
\item{x}{Vector of strings}

\item{char}{Vector of characters}
}
\description{
This function removes \code{char} from \code{x} and returns the new \code{x}. Vectorised over \code{x} and \code{char}.
}

