% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simpleCap}
\alias{simpleCap}
\title{Capitalize every first letter of a word}
\usage{
simpleCap(words)
}
\arguments{
\item{words}{a vector of strings}
}
\description{
Split the word at each space (\code{" "}) and capitalize every sub-word.
}

