% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_censor_data.R
\name{simulate_weibull_data}
\alias{simulate_weibull_data}
\title{Simulate Weibull censored data.}
\usage{
simulate_weibull_data(type, r, Pt, shape, scale)
}
\arguments{
\item{type}{Censoring Type}

\item{r}{Expected Number of Failures}

\item{Pt}{Fraction of Failing}

\item{shape}{The Weibull shape parameter}

\item{scale}{The Weibull scale parameter}
}
\description{
Simulate Weibull censored data.
}
\examples{
simulate_weibull_data(2, 10, 0.1, 1, 1)
}
