% Generated by roxygen2 (4.0.1): do not edit by hand
\name{split_by_prefix}
\alias{split_by_prefix}
\title{Split a data frame by prefix}
\usage{
split_by_prefix(df, prefix, keep = NULL)
}
\arguments{
\item{df}{a data frame}

\item{prefix}{character, prefix of the variables to be extracted}

\item{keep}{character vector, specify variables without \code{prefix} to be returned}
}
\value{
a \code{data.frame} with the variables renamed without the \code{prefix}.
}
\description{
Split a data frame by extracting only variables with a given \code{prefix}, allow for keeping some variables by passing their names in \code{keep}.
}

