% Generated by roxygen2 (4.0.2): do not edit by hand
\name{str_compare}
\alias{str_compare}
\title{Compare two strings}
\usage{
str_compare(a, b)
}
\arguments{
\item{a,b}{two strings}
}
\description{
Compare \code{a} and \code{b} after removing all digits, space and & characters.
Return \code{TRUE} if \code{a} is a sub-string of \code{b} (or vice-versa).
}

