% Generated by roxygen2 (4.0.2): do not edit by hand
\name{str_detect_multi_pattern}
\alias{str_detect_multi_pattern}
\title{Detect the presence or absence of pattern in a string}
\usage{
str_detect_multi_pattern(string, multi_pattern, expression = c("regexp",
  "fixed", "ignore.case", "perl"), value = FALSE)
}
\arguments{
\item{string}{input vector. This must be an atomic
  vector, and will be coerced to a character vector}

\item{multi_pattern}{vector of patterns to look for.}

\item{expression}{type of matching: POSIX regular expression (default), fixed, case insensitive and perl-compatible.}

\item{value}{logical, if \code{TRUE} then the values of matched elements are returned.}
}
\description{
This function does the same as \code{\link[stringr]{str_detect}} but allow for multiple pattern match and can return the values of matched elements.
}

