% Generated by roxygen2 (4.0.1): do not edit by hand
\name{write-and-reinstall-packages}
\alias{read_and_install_packages}
\alias{write-and-reinstall-packages}
\alias{write_installed_packages}
\title{Write and reinstall packages}
\usage{
write_installed_packages(dir = getwd())

read_and_install_packages(file)
}
\arguments{
\item{dir}{path of the directory to write the file (default to current working directory).}

\item{file}{path of the directory containing the file.}
}
\description{
These functions should be run immediately before and after upgrading \R as they allow to save all installed packages and reinstall them after upgrading.
\itemize{
	\item \code{write_installed_packages} write the name of all installed packages in a in a csv file "intalled_packages.csv".
	\item \code{read_and_install_packages} read the file and intall all packages with dependencies.
}
}

