% Generated by roxygen2 (4.0.2): do not edit by hand
\name{facet_wrap_labeller}
\alias{facet_wrap_labeller}
\title{Change strip texts of a facet wrap}
\usage{
facet_wrap_labeller(gplot, labels = NULL)
}
\arguments{
\item{gplot}{a \code{ggplot} or \code{grob} object.}

\item{labels}{character vector. Names of the strips. Strips are renamed in the same order as they are plotted, from left to right and top to bottom.}
}
\description{
This function allows to rename the strip of a facet wrap.
}

