% Generated by roxygen2 (4.0.1): do not edit by hand
\name{split_by_prefix}
\alias{split_by_prefix}
\title{Split a data frame using prefix}
\usage{
split_by_prefix(df, prefix, keep = NULL)
}
\arguments{
\item{df}{a data frame}

\item{prefix}{character, prefix of the variables to be extracted}

\item{keep}{character vector, specify variables without \code{prefix} to be returned}
}
\value{
a list of two elements:
\itemize{
	\item \code{df} original data frame without \code{prefix} variables
	\item \code{df_prefix} \code{data.frame} with \code{prefix} and \code{keep} variables renamed without the \code{prefix}.
}
}
\description{
Split by extracting variables with a given \code{prefix}. Keep variables without \code{prefix} by passing their names in \code{keep}.
}

